+++ src/Util/FilterUtil.php
@@ -1,0 +1,16 @@
+<?php
+namespace Zainpay\SDK\Util;
+
+class FilterUtil
+{
+    public static function ConstructFilterParams(?string $accountNumber, ?string $txnType,  ?string $dateFrom,  ?string $dateTo)
+
+    {
+        $params = [];
+        (!empty($accountNumber)) ? $params['accountNumber'] = $accountNumber : "";
+        (!empty($txnType))       ? $params['txnType'] = $txnType : "";
+        (!empty($dateFrom))      ? $params['dateFrom'] = $dateFrom : "";
+        (!empty($dateTo))        ? $params['dateTo'] = $dateTo : "";
+        return $params;
+    }
+}