+++ src/VirtualAccount.php
@@ -4,6 +4,7 @@ namespace Zainpay\SDK;
 
 use GuzzleHttp\Exception\GuzzleException;
 use Zainpay\SDK\Lib\RequestTrait;
+use Zainpay\SDK\Util\FilterUtil;
 
 class VirtualAccount
 {
@@ -30,9 +31,9 @@ class VirtualAccount
      * @throws GuzzleException
      * @link https://zainpay.ng/developers/api-endpoints?section=virtual-account-transactions
      */
-    public function transactionList(string $accountNumber, int $count = 20): Response
+    public function transactionList(string $accountNumber, int $count = 20, ?string $txnType,  ?string $dateFrom,  ?string $dateTo): Response
     {
-        return $this->get($this->getModeUrl() . 'virtual-account/wallet/transactions/' . $accountNumber . "/". $count);
+        return $this->get($this->getModeUrl() . 'virtual-account/wallet/transactions/' . $accountNumber . "/". $count, FilterUtil::ConstructFilterParams(null, $txnType, $dateFrom, $dateTo) );
     }
 
     /**