+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.7] - 2023-07-06
+## Changes
+- implemented filtering of transactions on txn endpoints
+
+## [1.0.6] - 2023-06-07
+## Changes
+- updated zainbox creation and updation payload to include allowAutoInternalTransfer
+
 ## [1.0.5] - 2023-05-31
 ## Changes
 - added optional parameter of setting count to endpoints for getting transactions list